1.19.3 update and fix to follow guidelines
updated the pack.mcmeta format to 12, added atlases files, and added the particle.json file for explosion particles allowing the duplicate particle textures to be removed. this also in theroy makes any particles applied by a different resourcepack appear on particles created by the datapack. Also added a 1 sample silent ogg file so that any vanilla explosion sounds do not send a warning to the log.